complete hw 20

+++ src/modules/blogger-platform/comments/application/usecases/delete-comment.usecase.ts
@@ -1,9 +1,13 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { CommentsRepository } from '../../infrastructure/comments.repository';
-import { NotFoundException } from '@nestjs/common';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
+import { CommentModelType } from '../../types/comment-model.type';
 
 export class DeleteCommentCommand {
-  constructor(public id: string) {}
+  constructor(
+    public id: string,
+    public userId: string,
+  ) {}
 }
 
 @CommandHandler(DeleteCommentCommand)
@@ -13,9 +17,14 @@ export class DeleteCommentUseCase
   constructor(private commentsRepository: CommentsRepository) {}
 
   async execute(command: DeleteCommentCommand) {
-    const comment = await this.commentsRepository.findComment(command.id);
+    const comment: CommentModelType | null =
+      await this.commentsRepository.findComment(command.id);
 
     if (!comment) throw new NotFoundException();
+    console.log(comment);
+    if (comment.userId !== command.userId) {
+      throw new ForbiddenException();
+    }
 
     await this.commentsRepository.deleteComment(command.id);
   }
